+++ src/PS2Challenge.Backend.Tests/Services/GameServiceTests.cs
@@ -11,6 +11,7 @@ public class GameServiceTests : IDisposable
 {
     private readonly Ps2ChallengeDbContext _context;
     private readonly GameService _gameService;
+    private bool _disposed;
 
     public GameServiceTests()
     {
@@ -21,8 +22,24 @@ public class GameServiceTests : IDisposable
 
     public void Dispose()
     {
-        _context.Database.EnsureDeleted();
-        _context.Dispose();
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (_disposed)
+        {
+            return;
+        }
+
+        if (disposing)
+        {
+            _context.Database.EnsureDeleted();
+            _context.Dispose();
+        }
+
+        _disposed = true;
     }
 
     [Fact]
